feat: 삭제 및 수정 API 일부 구현

+++ src/main/java/team7/hrbank/domain/Department/DepartmentRepository.java
@@ -60,5 +60,5 @@ public interface DepartmentRepository extends JpaRepository<Department, Long> {
                                                 Pageable pageable);
 
     //부서 아이디로 단건 조회
-    Department findById(long id);
+    Department findById(Integer id);
 }
